@@ -2,7 +2,8 @@
     "name": "vscode-church-slavonic-keyboard",
     "displayName": "vscode-church-slavonic-keyboard",
     "description": "Church Slavonic soft keyboard",
-    "version": "0.0.10",
+    "version": "0.1.0",
+    "type": "module",
     "repository": {
         "type": "git",
         "url": "https://github.com/slavonic/vscode-church-slavonic-keyboard"
@@ -10,7 +11,7 @@
     "publisher": "pgmmpk",
     "license": "MIT",
     "engines": {
-        "vscode": "^1.23.0"
+        "vscode": "^1.103.0"
     },
     "categories": [
         "Other"
@@ -18,7 +19,7 @@
     "activationEvents": [
         "*"
     ],
-    "main": "./extension",
+    "main": "./extension.js",
     "contributes": {
         "keybindings":
 [
@@ -3295,11 +3296,8 @@
         "test": "node ./node_modules/vscode/bin/test"
     },
     "devDependencies": {
-        "typescript": "^2.6.1",
-        "vscode": "^1.1.6",
-        "eslint": "^4.11.0",
-        "@types/node": "^7.0.43",
-        "@types/mocha": "^2.2.42",
-        "vsce": "^1.41.0"
+        "eslint": "^4.19.1",
+        "vsce": "^1.103.1",
+        "vscode": "^1.1.37"
     }
 }
